backend/requirements.txt

blinker==1.6.2
certifi==2022.12.7       
cffi==1.15.1
charset-normalizer==3.0.1
class-registry==2.1.2    
click==8.1.5
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.6
django-cors-headers==4.2.0
django-filter==22.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filters==1.3.2
filters-django==1.0.5
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
jwt==1.3.1
MarkupSafe==2.1.3
oauth2-provider==0.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.16.2